sriov: Refactor CmdAdd
Make the plugin/plugin.go call behave as a dispatcher, calling the right
CmdAdd function based on the device type is dealing with.
